drop the 'compliance' from the audit logs to make things a bit shorter
